Update use-metrics.html.md.erb
I made significant changes to this page now that we have a minor release past PCF 1.10+. This content was outdated as of PCF 1.10. Significantly trimmed down the content and redirected to new Monitoring PCF section, which is more helpful. Went ahead and left VM example now as I know people reference it, but made it a little more obvious it is examples, not comprehensive and not all ones we may or may not recommend